==> 3_programs/4_Parallel Trends - JG.Rmd <==
---
title: "Run analysis"
author: "Adelaide Currin, Jennah Gosciak"
date: "2022-11-05"
output: html_document
---

```{r, setup}
library(tidyverse)
library(rlang)
library(fs)
library(tidycensus)
library(lehdr)
library(sf)
library(readxl)
library(plm)
library(lmtest)
library(sandwich)
library(wfe)
data_dir <- "/Users/adelaide/Documents/GitHub/RMproject"

date <- Sys.Date() %>% 
  str_replace_all("-", "_")
```

```{r}
# define outcomes and covariates
outcomes <- c("C000rac", 
              "C000wac",
              "CE01rac", "CE01wac", "CE02rac", "CE02wac", "CE03rac", "CE03wac",
              "CR01rac", "CR01wac", "CR02rac", "CR02wac", "CRothrac", "CRothwac",
              "CT02rac", "CT02wac",
              "S000_same_tract", "S000_desig", "S000_elig")

log_outcomes <- str_c(outcomes, "_log")
pct_outcomes <- str_c(outcomes, "_pctchange")


```

```{r}
merged_oz <- readRDS("../2_output/analysis_file.RDS")

merged_oz %>% 
  group_by(year, DESIGNATED) %>% 
  summarize(n = n())


cov <- c("log_median_household_income", "total_housing", "percent_white", "percent_postsec", "percent_rent", "percent_hc_covered", "percent_poverty", "percent_suplemental_income", "percent_employed")

#only top commuting areas

topcommutezones <- merged_oz %>% 
  filter(`Our CBSA`==1)
```
```{r}
matched_data <- readRDS("../2_output/matched_analysis_file.RDS")
matched_data
```


# Plots

```{r}
generate_parallel_trends_plot <- function(df, type, outcome_type, df_type="all") {
  if (type == "log") {
    outcomes_list <- log_outcomes
  } else {
    outcomes_list <- pct_outcomes
  }
  
  if (outcome_type == "rac") {
    outcomes_list <- outcomes_list[str_detect(outcomes_list, "rac_")]
  } else if (outcome_type == "wac") {
    outcomes_list <- outcomes_list[str_detect(outcomes_list, "wac")]
  } else {
    outcomes_list <- outcomes_list[str_detect(outcomes_list, "same") |
                                     str_detect(outcomes_list, "elig") |
                                     str_detect(outcomes_list, "desig")]
  }
  df_pivot <- df %>% 
      pivot_longer(outcomes_list, names_to = "outcomes", values_to = "employment") %>% 
      filter(year > 2014) %>% 
      group_by(year, treatment, outcomes) %>% 
      mutate(employment = if_else(employment == Inf, 0, employment)) %>% 
      summarize(employment = mean(employment, na.rm = T))
  
  plot <- df_pivot %>% 
      ggplot(aes(x = year, y = employment)) +
      geom_line(aes(color=if_else(treatment == 1, "Selected", "Eligible"))) +
      #geom_smooth(aes(color=if_else(treatment == 1, "Selected", "Eligible")),
      #            method=lm) +
      theme_classic() +
      labs(color = "Designation") +
      facet_wrap(~outcomes, ncol=1, scales = "free")
  
  plot %>%
    print()
  
  ggsave(paste0("../2_output/", outcome_type, "_", type, "_",
  df_type, "_parallel_trends.png"),
         height = 15)
}
```
```{r, fig.height = 15, fig.width = 5}
# generate plots
generate_parallel_trends_plot(merged_oz, "pct", "rac")
generate_parallel_trends_plot(merged_oz, "pct", "wac")
generate_parallel_trends_plot(merged_oz, "pct", "od")

generate_parallel_trends_plot(topcommutezones, "pct", "rac", df_type="commute")
generate_parallel_trends_plot(topcommutezones, "pct", "wac", df_type="commute")
generate_parallel_trends_plot(topcommutezones, "pct", "od", df_type="commute")

generate_parallel_trends_plot(matched_data, "pct", "rac", df_type="matched")
generate_parallel_trends_plot(matched_data, "pct", "wac", df_type="matched")
generate_parallel_trends_plot(matched_data, "pct", "od", df_type="matched")
```

```{r, fig.height = 15, fig.width = 5}
# same as above, but dropping small tracts
generate_parallel_trends_plot(merged_oz %>% 
                                filter(C000rac > 1000), "pct")
```

```{r, fig.height = 15, fig.width = 5}
# same as above, but dropping small tracts
generate_parallel_trends_plot(merged_oz %>% 
                                filter(C000wac > 50), "pct")
```

```{r}
# functions for each model
pooled_ols <- function(outcome, data, covariates) {
  
  return(lm(paste0(outcome, " ~ treatment:year + year + treatment +  ", str_c(covariates, collapse = "+")),
               data = data))
}
```


```{r}
generate_model_output <- function(fit_model, outcome, treat_var, model_type, data, print) {
  sum_fit_plm <- summary(fit_model)
  coef_names <- sum_fit_plm$coefficients[, 1] %>% 
      names()
  coef_vals <- coeftest(fit_model, vcov = vcovCL,
                                    type = "HC1",
                                    cluster = data$GEOID)
  
  if (print == TRUE) {
      sum_fit_plm %>% 
      print() 
    
    coef_vals %>% 
      print()
  }
  
  coef_pos <- which(treat_var == coef_names)
  
  coef <- round(sum_fit_plm$coefficients[coef_pos, 1], 3)
  se <- round(coef_vals[coef_pos, 2], 3)
  pval <- round(coef_vals[coef_pos, 4], 4)
  coef_ast <- case_when(pval < 0.01 ~ paste0(coef, "***"),
                        pval < 0.05 ~ paste0(coef, "**"),
                        pval < 0.1 ~ paste0(coef, "*"))
  
  return(data.frame("outcome" = outcome,
               "treatment_post" = sum_fit_plm$coefficients[coef_pos, 1],
               "coef_asterisk" = coef_ast,
               "se" = se,
               "pval" = pval,
               "rsq" = round(sum_fit_plm$r.squared[[1]], 3),
               "model_type" = model_type))
}

run_model_outcome <- function(outcome, models, data, covariates=cov, print=TRUE) {
  covariates <- str_c(covariates, collapse = " + ")
  print(str_c("Running model for ", outcome))
  results <- data.frame()
  
  if ("did" %in% models) {
    fit_did <- pooled_ols(outcome, data, covariates)
    results <- bind_rows(results, generate_model_output(fit_did, outcome,
                                                        "treatment:year", "did", data, print=print)) 
  }
 
  return(results)
}

models <- c("did")
run_model_outcome("S000_same_tract", models, data = merged_oz %>% 
                    filter(post == 0), print=T)
run_model_outcome("S000_same_tract", models, data = topcommutezones %>% 
                    filter(post == 0), print=T)
```
```{r}
results <- map_dfr(c(log_outcomes), ~run_model_outcome(., models=models, 
                                                                 data = merged_oz %>% 
                                                                   filter(post == 0), print=F))
results
```

```{r}
rac_log_outcomes <- log_outcomes[str_detect(log_outcomes, "rac")]
wac_log_outcomes <- log_outcomes[str_detect(log_outcomes, "wac")]

rac_pct_outcomes <- pct_outcomes[str_detect(pct_outcomes, "rac")]
wac_pct_outcomes <- pct_outcomes[str_detect(pct_outcomes, "wac")]

# drop tracts with low employment
merged_oz %>% 
  summarize(rac_p5 = quantile(C000rac, 0.05),
            wac_p5 = quantile(C000wac, 0.05))
```

```{r}
results_subsample_rac <- map_dfr(c(rac_log_outcomes), ~run_model_outcome(., models=models, 
                                                                 data = merged_oz %>% 
                                                                   filter(post == 0 & C000rac > 500), print=F))
results_subsample_rac

results_subsample_wac <- map_dfr(c(wac_log_outcomes), ~run_model_outcome(., models=models, 
                                                                 data = merged_oz %>% 
                                                                   filter(post == 0 & C000wac > 50), print=F))
results_subsample_wac
```

```{r}
resultspct <- map_dfr(pct_outcomes, ~run_model_outcome(., models=models, 
                                                                 data = merged_oz %>% 
                                                                 filter(year > 2014), print=F))
resultspct
```

```{r}
resultspctcommuting <- map_dfr(pct_outcomes, ~run_model_outcome(., models=models, 
                                                                 data = topcommutezones %>% 
                                                                 filter(year > 2014), print=F))
resultspctcommuting
```


```{r}
ols_balance <- function(outcome, data) {
  return(lm(paste0(outcome, " ~ treatment"),
               data = data))
}

run_balance_tests <- function(outcome, data, print=TRUE) {
  print(str_c("Running model for ", outcome))
  results <- data.frame()
  
  if ("did" %in% models) {
    fit_did <- ols_balance(outcome, data)
    results <- bind_rows(results, generate_model_output(fit_did, outcome,
                                                        "treatment", "did", data, print=print)) 
  }


}

# balance test
balance_tests <- map_dfr(c(cov), ~run_balance_tests(., merged_oz %>%
                                                      filter(year == 2017)))
balance_tests

balance_tests_commuting <- map_dfr(c(cov), ~run_balance_tests(., topcommutezones %>%
                                                      filter(year == 2017)))
balance_tests_commuting
```
