==> 3_programs/5_Weighting_JG_testing.Rmd <==
---
title: "Run analysis"
author: "Adelaide Currin, Jennah Gosciak"
date: "2022-11-05"
output: html_document
---

```{r, setup}
library(tidyverse)
library(rlang)
library(fs)
library(tidycensus)
library(lehdr)
library(sf)
library(readxl)
library(plm)
library(lmtest)
library(sandwich)
library(wfe)
library("MatchIt")
library("optmatch")
library("marginaleffects")
library(WeightIt)
library(cobalt)
library(survey)

date <- Sys.Date() %>%
  str_replace_all("-", "_")
```

```{r}
# define outcomes and covariates
outcomes <- c(
  "C000rac",
  "C000wac",
  "CE01rac", "CE01wac", "CE02rac", "CE02wac", "CE03rac", "CE03wac",
  "CR01rac", "CR01wac", "CR02rac", "CR02wac", "CRothrac", "CRothwac",
  "CT02rac", "CT02wac",
  "S000_same_tract", "S000_desig", "S000_elig"
)

log_outcomes <- str_c(outcomes, "_log")
pct_outcomes <- str_c(outcomes, "_pctchange")
cov <- c("log_median_household_income", "total_housing", "percent_white", "percent_postsec", "percent_rent", "percent_hc_covered", "percent_poverty", "percent_suplemental_income", "percent_employed", "dec_score", "SE_Flag", "vacancyrate", "pctunder18", "pctover64")
```

```{r}
merged_oz <- readRDS("../2_output/analysis_file.RDS") %>%
  mutate(
    year_fact2014 = if_else(year == 2014, 1, 0),
    year_fact2015 = if_else(year == 2015, 1, 0),
    year_fact2016 = if_else(year == 2016, 1, 0),
    year_fact2017 = if_else(year == 2017, 1, 0),
    year_fact2018 = if_else(year == 2018, 1, 0),
    year_fact2019 = if_else(year == 2019, 1, 0)
  ) %>%
  mutate(across(cov, ~ if_else(is.na(.), 0, .)))
merged_oz

merged_oz_weight_sample <- merged_oz %>%
  filter(`Our CBSA` == 1 & year == 2017)
```

```{r}
covariates <- str_c(cov, collapse = " + ")
form <- formula(paste0("treatment ~", covariates))
# m.out0 <- matchit(form, data = merged_oz %>%
#                     filter(year == 2017 & `Our CBSA`==1),
#                  method = NULL, distance = "glm")
# m.out0
#
# summary(m.out0)

bal.tab(form,
  data = merged_oz_weight_sample, estimand = "ATT", thresholds = c(m = .05)
)
```
```{r}
# weighting?
W.out <- weightit(form,
  data = merged_oz %>%
    filter(year == 2017 & `Our CBSA` == 1), estimand = "ATT", method = "ps"
)
W.out

summary(W.out)
```

```{r}
bal.tab(W.out, stats = c("m", "v"), thresholds = c(m = .05))
```

```{r}
merged_oz_weight_sample$weights <- W.out$weights

merged_oz_wgt <- merged_oz %>%
  filter(`Our CBSA` == 1) %>%
  left_join(merged_oz_weight_sample %>%
    select(GEOID, weights), by = "GEOID")

stopifnot(sum(merged_oz_wgt$weights == 0) == 0)
```


```{r}
# matching
# m.out1 <- matchit(form, data = merged_oz %>%
#                     filter(year == 2017 & `Our CBSA`==1),
#                   method = "nearest", distance = "glm")
```

```{r}
# summary(m.out1, un = FALSE)
```
```{r}
# plot(m.out1, type = "jitter", interactive = FALSE)
```

```{r}
# plot(m.out1, type = "density", interactive = FALSE,
#      which.xs = c(cov))
```
```{r}
# plot(summary(m.out1))
```
```{r}
# m.data <- match.data(m.out1)
#
# m.data_panel <- inner_join(merged_oz,
#                           m.data %>%
#                             select(GEOID, subclass, weights, distance),
#                           using = "GEOID")
# head(m.data_panel)
# m.data_panel %>%
#   saveRDS("../2_output/matched_analysis_file.RDS")
```
```{r}
# functions for each model
pooled_ols <- function(outcome, data, covariates) {
  return(lm(
    formula(paste0(
      outcome, " ~ post_treatment + ",
      "treatment + post + year_fact + ", covariates
    )),
    data = data,
    weights = weights
  ))
}

trend_model <- function(outcome, data, covariates) {
  return(lm(
    paste0(
      outcome, " ~ treatment:year:post + treatment:post + treatment:year +",
      "treatment + year + post +", covariates
    ),
    data = data, weights = weights
  ))
}


twfe_model <- function(outcome, data, covariates) {
  return(plm(formula(paste0(outcome, " ~ post_treatment")),
    data = data,
    index = c("GEOID", "year"),
    model = "within",
    effect = "twoways", weights = weights
  ))
}

gen_covariates_int <- function(covariates, year_vars) {
  covariates_list <- map(year_vars, ~ str_c(covariates, ":", .)) %>%
    unlist() %>%
    str_c(collapse = " + ")
  return(covariates_list)
}

twfe_model_covariates <- function(outcome, data, covariates) {
  covariates_list <- gen_covariates_int(covariates, str_c("year_fact", c(2014:2019)))
  return(plm(paste0(outcome, " ~ post_treatment + ", covariates_list),
    data = data,
    index = c("GEOID"),
    model = "within",
    effect = "individual", weights = weights
  ))
}

event_model <- function(outcome, data) {
  mod <- plm(
    paste0(
      outcome, " ~ treatment:year_fact2014 + treatment:year_fact2015 + ",
      "treatment:year_fact2016 + treatment:year_fact2018 + treatment:year_fact2019 + ",
      "year_fact2014 + year_fact2015 + year_fact2016 + year_fact2018 + year_fact2019"
    ),
    data = data,
    index = c("GEOID"),
    model = "within",
    effect = "individual",
    weights = weights
  )

  return(mod)
}
```

```{r}
generate_model_output <- function(fit_model, outcome, treat_var, model_type, data, print) {
  sum_fit_plm <- summary(fit_model)
  coef_vals <- coeftest(fit_model,
    vcov = vcovCL,
    type = "HC1",
    weights = data$weights,
    cluster = data$GEOID
  )

  if (print == TRUE) {
    print(str_c("Running ", model_type, " model for ", outcome))

    coef_vals %>%
      print()
  }

  coef <- round(coef_vals[2, 1], 3)
  se <- round(coef_vals[2, 2], 3)
  pval <- round(coef_vals[2, 4], 4)
  coef_ast <- case_when(
    pval < 0.01 ~ paste0(coef, "***"),
    pval < 0.05 ~ paste0(coef, "**"),
    pval < 0.1 ~ paste0(coef, "*")
  )

  return(data.frame(
    "outcome" = outcome,
    "treatment_post" = coef,
    "coef_asterisk" = coef_ast,
    "se" = se,
    "pval" = pval,
    "rsq" = round(sum_fit_plm$r.squared[[1]], 3),
    "model_type" = model_type
  ))
}

run_model_outcome <- function(outcome, models, data = merged_oz_wgt,
                              covariates = cov, print = TRUE) {
  covariates_orig <- covariates
  covariates <- str_c(covariates, collapse = " + ")
  results <- data.frame()
  if ("twfe" %in% models) {
    fit_plm <- twfe_model(outcome, data, covariates)
    results <- bind_rows(results, generate_model_output(fit_plm, outcome,
      "post_treatment", "twfe", data,
      print = print
    ))
  }
  if ("twfe_cov" %in% models) {
    fit_plm <- twfe_model_covariates(outcome, data, covariates_orig)
    results <- bind_rows(results, generate_model_output(fit_plm, outcome,
      "post_treatment", "twfe_cov", data,
      print = print
    ))
  }
  if ("did" %in% models) {
    fit_did <- pooled_ols(outcome, data, covariates)
    results <- bind_rows(results, generate_model_output(fit_did, outcome,
      "post_treatment", "did", data,
      print = print
    ))
  }
  if ("time_trend" %in% models) {
    fit_trend_lm <- trend_model(outcome, data, covariates)
    results <- bind_rows(results, generate_model_output(fit_trend_lm, outcome,
      "treatment:year:post", "time_trend", data,
      print = print
    ))
  }
  if ("event" %in% models) {
    fit_trend_lm <- event_model(outcome, data)
    results <- bind_rows(results, generate_model_output(fit_trend_lm, outcome,
      "treatment:year_fact2019", "event", data,
      print = print
    ))
  }
  return(results)
}

models <- c("did")
run_model_outcome("C000rac_pctchange", models, print = T)
```

```{r}
# results <- map_dfr(c(log_outcomes), ~run_model_outcome(., models=models, print=FALSE))
# results
#
# # save to output
# results %>%
#   write_csv(str_glue("../2_output/results_log_ps_{date}.csv"), na="")
```

```{r}
results <- map_dfr(c(pct_outcomes), ~ run_model_outcome(.,
  models = c("did"),
  data = merged_oz_wgt %>%
    filter(year > 2014), print = F
))
results

# save to output
# results %>%
#   write_csv(str_glue("../2_output/results_pct_ps_{date}.csv"), na="")
```
```{r}
# placebo test
placebo_df <- merged_oz_wgt %>%
  filter(post == 0) %>%
  select(-c(post, post_treatment)) %>%
  mutate(
    post = if_else(year >= 2016, 1, 0),
    post_treatment = post * treatment
  )

# placebo_results <- map_dfr(c(log_outcomes), ~run_model_outcome(., models=c("did"),
#                                                                data = placebo_df,
#                                                     print=FALSE))
# placebo_results
#
# # save to output
# placebo_results %>%
#   write_csv(str_glue("../2_output/placebo_log_ps_{date}.csv"), na="")
```
```{r}
# placebo test
placebo_results <- map_dfr(c(pct_outcomes), ~ run_model_outcome(.,
  models = c("did"),
  data = placebo_df %>%
    filter(year > 2014),
  print = FALSE
))
placebo_results

# save to output
placebo_results %>%
  write_csv(str_glue("../2_output/placebo_pct_ps_{date}.csv"), na = "")
```
